+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Make `UploadedFile` actually delete itself only once (janko-m)
+
 * Make `UploadedFile#inspect` cleaner by showing only the data hash (janko-m)
 
 * Make determine_mime_type able to accept non-files when using :file (janko-m)